refactor: replace aspect ratio icon logic with a standalone pipe and update aspect ratio state initialization
